+++ src/SME.SR.Workers.SGP/Views/HistoricoEscolar/DadosHistoricoEJA.cshtml
@@ -82,7 +82,7 @@
 
     string RetornarTipoNotaFrequencia(string tipoNotaConceito)
     {
-        return $"<td class={"texto-centralizado coluna-nota"}>{tipoNotaConceito}</td> <td class={"texto-centralizado coluna-nota"}>F</td>";
+        return $"<td class='texto-centralizado coluna-nota'>{tipoNotaConceito}</td> <td class='texto-centralizado coluna-nota'>F</td>";
     }
     string ObterLinhasBaseNacional()
     {
